requirements.txt
paho-mqtt>=2.0.0
openai>=1.0.0
sounddevice>=0.4.6
numpy>=1.24.0
pyyaml>=6.0
torch>=2.0.0
silero-vad>=5.0
